+++ packages/ragbits-agents/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- add Context to Agents (#715)
+
 ## 1.1.0 (2025-07-09)
 
 ### Changed